Unity Project
Initial Unity project with default .gitignore from GitHub. No changes have been made for the project yet; this is just to set things up. As long as it's the 2020 version, it should be fine, as it doesn't need to be updated. The version it's running on is 2020.1.6f1, and shouldn't need to be changed even if Unity comes out with a new update.