Added Timer DLLs
Added the stopwatch and countdown timer DLLs. We still need DLLs for the different design patterns, but we at least have something to show that we know DLLs work. The Timer.cs file that was already in place is still there, just moved into a different folder, and disabled on the text object. The text object instead has a TimeText.cs script enabled, which allows for you to switch between a countdown timer or stopwatch timer.