Improved Jump*
The jump has been fixed so that it's actually consistent and at a good amount. However, due to a problem with ground detection the player can do a jump in the air if they did not jump to get into air to start with. But I'm going to leave that behind for now.