Added Moving Platform Script
Added a script that allows for moving platforms. It requires a rigid body, but if you don't add one it a rigid body will be added automatically. If you don't want the platform to tilt, freeze the rotations on the rigid body. If you don't want the platform to sink when weight is applied, increase its mass to withstand said weight.