Fixing EXE Problem + Resources Folder
The EXE could not build because of "using UnityEditor" and "using TreeEditor". It appears that any inclusion that has "Editor" in its name can only work in the Unity project, and not the EXE. So this content was taken out.

In order to load objects from a file path (which is what the "using UnityEditor" was for), we now use Resources.Load(). However, this requires that said content is in the "Assets/Resources" folder, so I moved all but the Scripts, Plugins, and Scenes in there. Basically anything that wouldn't go in your standard "res" folder was put in here. The GameplayManager can generate players at runtime, but it may not be best to do it like that.

Finally, I added a prefab for a moving platform. I'll try to create prefabs for the animal players as well.