Going to Limit Record Sizes
So as it turns out, due to size issues the records aren't saving properly. So I'm going to have to put a limit on them.
One file will be used to track the amount of lines an item takes up, the next one will be used to hold the data itself.