Adding in my Tutorial
In order to meet the requirements I had to add in my tutorial. It is accessible via the "tutorial 2" button. I also made slight changes and added a folder for the profile data.

I need to make more changes, but I won't do them before the presentation. So I'll leave them for later and work on other things.