Working on Flag Prefab
I changed the flag prefab so that it uses the flag object I provided. For changing the flag object, you just need to add a new object as a child of the flag object, give it a box collider and FlagObject script, and turn off the other representative flag objects under the same parent. I also made a basket prefab for the Halloween stage, though I haven't replaced the flag indicator yet.