Fixed Christmas Stage + Note About Project Updates
I needed to fix the models so that the transformation information is saved to the game and not to the model files themselves. Basically the transformations were not applied before the models were exported, so I'm fixing that now. The Christmas stage has been fixed, so now I'm going to do the Halloween stage.

As a side note, I'm going to close down the DLL specific repo and just do the work in the main game project repo instead. This will happen on January 5th at the latest, though it will definitely be sometime after New Years.

Finally, I'm using this branch because the main branch still might be getting looked at for a past submission. When I get a note back about the process I'll merge this with the main branch. Other than that, this branch will probably just be used for testing more unstable, uncertain game features.