Stage Save Issue Resolved
So, for some reason if the position values are too high (I'm guessing this goes for rotation an scale as well), the item will save but not load properly. So when it comes to the parent object that holds the parent object for the stage items, make sure its positioned at (0, 0, 0), its scale is (1, 1, 1), and its rotation is (0, 0, 0, 1).