Changing GameBuilder (WIP)
The GameBuilder has been fixed (in part). Basically since the game returns to the main menu after a match ends, the game builder kept getting remade. As such, I now destroy the game builder when a game ends.

The replay button needs to be fixed (it leads to the empty scene), and I need to setup the game builder for the tutorials as well.