Jump Is Mostly Fixed, Now Revamping Tutorial
The jump is mostly fixed. I have one hitbox for jumping and one for general collisions. This doesn't fix the player hanging onto surfaces when holding forward problem, but it fixes the other issues. The tutorials need revamping now to use the new level loading system, which I'll do now.