Working on Servers (WIP)
Still working on servers. 4 player server works, but the program would be slowed down if it constantly checked with data for all four sockets, so I'm going to change it to using lists instead.

TODO: provide option to delete existing array data when setting array (see Array.Clear()), or simply replace its contents. Array.Clear deletes the data from memory, which the user may not want to do since they might still have a reference to it.

TODO: provide option to add or remove endpoints to listen to for UdpServerX.