Working on Online
Working on the online functionality for the game. The setup for applying and getting data of the data transfers is mostly setup, so I should be able to test things soon. I need to think about the item boxes will be handled, as I need a way to check which items have been used up (maybe have item box numbers?), but I should hopefully be good to get things working by the end of today.