Going to Test Online Soon
I'm going to test out the online for the game soon. In order to do that I'll need to build a debug EXE for the server test and client test, which I'll do soon. This EXE build was just to update the current EXE with what we have.

Also, I added in some things for the Valentine's Day stage. I haven't tested it, so it might not be best to run it right now, but it's there. On a different note, I'm planning to just switch to a player list for the game, which might screw some things up. 

NOTE TO SELF: I haven't done it yet, but it's something I'll be looking at, so I'll need to see if this causes problems with the game builder or not. Maybe I should just leave it.