4 Player Connections - Model Inconsistencies
The 4 connections work now, but there's an issue. The models don't match up consistently, so anything above two players when multiple slots are available causes the data to get set to multiple players. This is a major issue, but things are still functional. This might be left on the back burner while I work on other online components.