Added Game End by Timeout
Added the ability for the game to end if time reaches 0, though this is disabled in all levels. I also added a CTRL + SHIFT + D shortcut to enter the debug level on the main menu. The debug button should be taken out at some point.

The EXE has been rebuilt as well.