Reworked Follower Camera Setup
I've re-worked the setup of the camera, which should fix the problem where you spawn without a camera attached. 

I still need to change how the create player process works for going back and fourth between the character menu, though. I'm thinking of changing where the player is physically created, and having the list clear when going back to the previous menu.