Created Game Settings
The game settings have been made. They're a little bit finicky, but they work for now. They should be revisited though. There's a bool attached to the SettingInterface script for controlling whether or not to load in the volume settings from the file to start off. There's also a copy with the default volume settings, where-in which the BGM volume is set to 0.2.