Game Start on Client Side is Broken
The game start on the client side no longer works, so changes need to be made.  The server name also doesn't seem to show up on the host screen anymore, so that needs to be fixed. Also, the host button should be taken away when someone enters in to join. The vice versa should also happen.