Fixed EXE problem.
Turns out I had "Create Visual Studio Solution" toggled on in the build settings, so it ended up not building the EXE. As for how it stands now, for some reason the fourth player does not join properly, and I don't know why. I think I'll try to fix the game start issue for now and come back to that later. I also need to account for the room being bigger than the amount of people it accounts for.