Pushing Current Progress
Still trying to figure out why the component gets deleted. Maybe it has something to do with destroying the player that has been saved when a new player is created. Need to find where that is.

Also, the game might stall if you close the client window before the server window when in-game. You need to fix that. Also, it appears you need to add a function to refresh the online game start to get rid of the deleted player in the list.

It should also be noted that the manager gets P0 but not P1, for some reason.