Positioning Fixed on Load Up
The positioning for the players works now. However, the client side does not appear to be updating with the movements from the server. I also need to fix the issue where the online lobby manager still holds the null value for the deleted debug character. Also, going back to the lobby doesn't work. When you go back, it just snaps back to the gameplay screen, except without the other players. Keep in mind that I haven't tested with more than 2 players (i.e. host and client).