Current Progress and Future Plans.
Thanks to the timeout variables, the games no longer stop when all the clients disconnect for the server. However, timing is still something that needs to be accounted for. 

The data still isn't sending properly to more than two players. You set up a file to use intervals, which should be looked at to try and help with this, as right now packets are sent every frame.

A file exists for controlling the HUD and updating it with game values. However, nothing is in it yet. So that's something to look it.

You should probably try to use online prediction to handle the players so that things are more synced up. Also, maybe add another line to see who has the flag just to make sure the worlds match up better.

On a different note, make sure the item icons don't show up if it's not the local player that has the item.