Intervals are Setup for Lobby - Didn't Fix Anything
As it turns out, not all of the data is being sent on either the lobby or manager side. I don't know why this is, but it is. I don't know how to fix it, and honestly I'm not sure if I'm even going to try. I'm probably going to just setup the HUD, maybe the other items, and maybe take a hand at online prediction. Will probably setup the intervals for the gameplay manager as well.