Going to Fix Christmas Stage
As it turns out, I'm going to backtrack on those plans. The lobby is more or less fixed, but the Christmas stage doesn't work (it's the only stage that does). The data isn't transferring properly, but I don't know what to do about that. And quite honestly, I think I'll just move onto other projects.