Christmas Stage Works Now
As it turns out, the Christmas stage internally was named "Christmasmap" still, even though I changed the name to "ChristmasMap". I fixed this by completely changing the scene name and then changing it to what I actually wanted. I might actually be able to setup a system where I'm able to go back into the lobby, but I'm going to keep working on that.