Multiplayer Hud Setup, Not Implemented Yet.
The multiplayer hud is setup, but it hasn't been tested yet. Uncomment the code to instantiate the HUD object, and put the existing HUD objects in the debug game object so that they get deleted.

On a side note, there appears to be something wrong with how you're reading the data from the array. Try testing it out outside of the program, because as it stands it appears that you're not moving the index through properly. Also see if you're meant to change System.Text.Encoding.ASCII.GetString() by replacing ASCII with UTF8 instead, since that's what you're using to encode in the first place.