Replaced HUD. Player icons and timer need work.
The HUD has been replaced. Now we have a HUD for multiplayer and single player. I need to have a function to return character type since the casting isn't working for choosing the right icon. Oh, and I added in the rest of the items too, though the icons need to be reworked. The EXE has not been rebuilt yet.

Still need to fix the data reading and writing.  After that, I want to take a shot at dead reckoning as the final thing. Also, I need to display the player scores in the menu, and setup the countdown timer.