Going to Fix Bunny and Turtle
So as it turns out, the bunny and turtle aren't actually using the right scripts... I need to fix that. Also, I found out that ASICII conversion isn't being used, and that with ASCII conversion each char is one byte, even though sizeof() says char is equal to 2. So that's something to fix. Also going to need to rebuilt the EXE.